% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anevah.R
\name{anevah}
\alias{anevah}
\title{ANEVA-H}
\usage{
anevah(
  ref_table,
  alt_table,
  min_occurances = 10L,
  min_cases = 6L,
  min_expr = 30L
)
}
\arguments{
\item{ref_table}{Data frame of reference counts, see below for more details}

\item{alt_table}{Data frame of alternate counts, see below for more details}

\item{min_occurances}{Minimum number of people who contain the gene}

\item{min_cases}{Minimum number of samples for gene expression}

\item{min_expr}{Minimum expression of a gene}
}
\value{
A data frame with two columns:
\describe{
 \item{\dQuote{GENE_ID}}{The gene ID}
 \item{\dQuote{Vg}}{The Vg estimate}
}
}
\description{
Fits data to BLN distribution using likelihood function, then estimates Vgs
}
\details{
The reference and alternate counts tables should be data frames. The rows
should be genes (must be named) and the columns should be samples (names
optional). For example:
\preformatted{
        Sample1    Sample2
gene1        0          3
gene2        1          2
}
\code{ref_table} and \code{alt_table} must have the same dimensions
}
