% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln.R
\name{bln_likelihood}
\alias{bln_likelihood}
\title{BLN Likelihood}
\usage{
bln_likelihood(ref_counts, total_counts, sd)
}
\arguments{
\item{ref_counts}{Reference counts}

\item{total_counts}{Total counts}

\item{sd}{A vector of standard deviations}
}
\value{
The negative log-likelihood
}
\description{
Calculate the negative total log-likelihood from a Binomial-Logit-Normal
(BLN) distribution
}
\keyword{internal}
