% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{estimate_vg}
\alias{estimate_vg}
\title{Calculate Vg Estimates}
\usage{
estimate_vg(sigma)
}
\arguments{
\item{sigma}{Standard deviation}
}
\value{
Vg estimates
}
\description{
Calculate Vg Estimates
}
\keyword{internal}
