% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln.R
\name{fit_bln}
\alias{fit_bln}
\title{Fit BLN}
\usage{
fit_bln(ref_counts, total_counts)
}
\arguments{
\item{ref_counts}{Reference counts}

\item{total_counts}{Total counts}
}
\value{
Returns the estimate of the standard deviation of our BLN data (sigma)
}
\description{
Optimizer for finding sigma
}
\keyword{internal}
