% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{formula_expectation}
\alias{formula_expectation}
\title{Expectation Formula}
\usage{
formula_expectation(sigma)
}
\arguments{
\item{sigma}{Standard deviation}
}
\value{
A formula function
}
\description{
Expectation of \eqn{ln(e)} to be used in \code{\link{formula_vg}}
}
\keyword{internal}
