% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{p_sr}
\alias{p_sr}
\title{Probability of SR}
\usage{
p_sr(sigma, sr)
}
\arguments{
\item{sigma}{Standard deviation}

\item{sr}{Residual of regulatory variance}
}
\value{
...
}
\description{
Probability of SR
}
\keyword{internal}
